refactor(utils): moved compared_authors_or_editors to dqm_ingest specific utils
